backend/requirements.txt
# Flask and dependencies
Flask==2.2.3
Flask-Cors==3.0.10

# PyTorch (with CUDA support for GPU acceleration, replace cu118 with your CUDA version)
torch==2.0.1+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.2+cu118
-f https://download.pytorch.org/whl/torch_stable.html

# Hugging Face Diffusers and related tools
diffusers==0.14.0
transformers==4.26.1
safetensors==0.3.1  # For efficient model handling

# Image processing
Pillow==9.4.0

# Logging and utilities
loguru==0.6.0
python-dotenv==1.0.0  # For environment variable management

# Optional memory optimizations
xformers==0.0.20  # Only install if needed for efficient GPU usage (optional)

# Testing and development tools
pytest==7.4.0
pytest-cov==4.1.0
#python -m venv venv